Implementações

+++ src/main/java/br/edu/utfpr/dungeontable/service/CampaignService.java
@@ -1,5 +1,7 @@
 package br.edu.utfpr.dungeontable.service;
 
+import br.edu.utfpr.dungeontable.exception.BusinessException;
+import br.edu.utfpr.dungeontable.exception.ErrorCode;
 import br.edu.utfpr.dungeontable.model.table.Campaign;
 import br.edu.utfpr.dungeontable.repository.CampaignRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,11 +18,19 @@ public class CampaignService {
 
     @Transactional(propagation = Propagation.REQUIRED)
     public Campaign save(Campaign campaign) {
+        if (campaign.getName() == null || campaign.getName().isEmpty()) {
+            throw new BusinessException(ErrorCode.ATTRIBUTE_REQUIRED, "name campaign");
+        }
         return campaignRepository.save(campaign);
     }
 
     @Transactional(propagation = Propagation.REQUIRED)
     public Campaign update(Campaign campaign) {
+        if(campaign.getId() == null){
+            throw new BusinessException(ErrorCode.ID_REQUIRED);
+        }else if (campaign.getName() == null || campaign.getName().isEmpty()) {
+            throw new BusinessException(ErrorCode.ATTRIBUTE_REQUIRED, "name campaign");
+        }
         return campaignRepository.save(campaign);
     }
 
@@ -34,6 +44,9 @@ public class CampaignService {
 
     @Transactional(propagation = Propagation.REQUIRED)
     public void delete(Long id) {
+        if(id == null){
+            throw new BusinessException(ErrorCode.ID_REQUIRED);
+        }
         campaignRepository.deleteById(id);
     }
 }
