+++ src/main/java/br/edu/utfpr/dungeontable/service/PlayerService.java
@@ -1,6 +1,8 @@
 package br.edu.utfpr.dungeontable.service;
 
 
+import br.edu.utfpr.dungeontable.exception.BusinessException;
+import br.edu.utfpr.dungeontable.exception.ErrorCode;
 import br.edu.utfpr.dungeontable.model.table.Player;
 import br.edu.utfpr.dungeontable.repository.PlayerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,12 +18,45 @@ public class PlayerService {
     private PlayerRepository playerRepository;
 
     @Transactional(propagation = Propagation.REQUIRED)
-    public Player save(Player player) {
+    public Player save(Player player) throws BusinessException {
+        if (player.getNamePlayer() == null || player.getNamePlayer().isEmpty()) {
+            throw new BusinessException(ErrorCode.ATTRIBUTE_REQUIRED, "name player");
+        }else if (player.getNameCharacter() == null || player.getNameCharacter().isEmpty()) {
+            throw new BusinessException(ErrorCode.ATTRIBUTE_REQUIRED, "name character");
+        } else if (player.getAntecedent() == null || player.getAntecedent().isEmpty()) {
+            throw new BusinessException(ErrorCode.ATTRIBUTE_REQUIRED, "antecedent");
+        } else if (player.getRaceCharacter() == null || player.getRaceCharacter().isEmpty()) {
+            throw new BusinessException(ErrorCode.ATTRIBUTE_REQUIRED, "race character");
+        } else if (player.getAge() == null) {
+            throw new BusinessException(ErrorCode.ATTRIBUTE_REQUIRED, "age");
+        } else if (player.getUser() == null) {
+            throw new BusinessException(ErrorCode.ATTRIBUTE_REQUIRED, "user");
+        } else if (player.getClassPlayer() == null) {
+            throw new BusinessException(ErrorCode.ATTRIBUTE_REQUIRED, "Class Player");
+        }
         return playerRepository.save(player);
     }
 
     @Transactional(propagation = Propagation.REQUIRED)
     public Player update(Player player) {
+        if(player.getId() == null){
+            throw new BusinessException(ErrorCode.ID_REQUIRED);
+        }
+        else if (player.getNamePlayer() == null || player.getNamePlayer().isEmpty()) {
+            throw new BusinessException(ErrorCode.ATTRIBUTE_REQUIRED, "name player");
+        }else if (player.getNameCharacter() == null || player.getNameCharacter().isEmpty()) {
+            throw new BusinessException(ErrorCode.ATTRIBUTE_REQUIRED, "name character");
+        } else if (player.getAntecedent() == null || player.getAntecedent().isEmpty()) {
+            throw new BusinessException(ErrorCode.ATTRIBUTE_REQUIRED, "antecedent");
+        } else if (player.getRaceCharacter() == null || player.getRaceCharacter().isEmpty()) {
+            throw new BusinessException(ErrorCode.ATTRIBUTE_REQUIRED, "race character");
+        } else if (player.getAge() == null) {
+            throw new BusinessException(ErrorCode.ATTRIBUTE_REQUIRED, "age");
+        } else if (player.getUser() == null) {
+            throw new BusinessException(ErrorCode.ATTRIBUTE_REQUIRED, "user");
+        } else if (player.getClassPlayer() == null) {
+            throw new BusinessException(ErrorCode.ATTRIBUTE_REQUIRED, "Class Player");
+        }
         return playerRepository.save(player);
     }
 
@@ -35,6 +70,9 @@ public class PlayerService {
 
     @Transactional(propagation = Propagation.REQUIRED)
     public void delete(Long id) {
+        if(id == null){
+            throw new BusinessException(ErrorCode.ID_REQUIRED);
+        }
         playerRepository.deleteById(id);
     }
 }
